requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.69
esprima==4.0.1
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
librecaptcha==0.7.3
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
pydub==0.25.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.3
requests==2.28.2
requests-toolbelt==0.10.1
selenium==4.8.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.9.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
websockets==10.4
wsproto==1.2.0
xvfbwrapper==0.2.9
yarl==1.8.2
